+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 ### Changed
 
+- Drop `once_cell` dependency by replacing `once_cell::sync::Lazy` with `std::sync::LazyLock`
+  - Needs `rust 1.80`
 - Make types related to full calc results `Deserialize`
 - Make types related to anagrafica `Deserialize`
 - Moved templates data for tests `from tests::engines::{hfbi, niseci}::full` to `tests::test_utils`